requirements.txt
CacheControl==0.14.2
Jinja2==3.1.4
MarkupSafe==3.0.1
PyYAML==6.0.2
Pygments==2.19.1
RapidFuzz==3.12.1
aiohappyeyeballs==2.4.3
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.8
cohere==5.13.12
comm==0.2.2
crashtest==0.4.1
debugpy==1.8.11
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
dulwich==0.22.7
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.8
fastavro==1.9.7
fastjsonschema==2.21.1
filelock==3.16.1
findpython==0.6.2
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.21.0
google-api-python-client==2.161.0
google-auth-httplib2==0.2.0
google-auth==2.35.0
google-generativeai==0.8.4
google==3.0.0
googleapis-common-protos==1.65.0
grpcio-status==1.62.3
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx-sse==0.4.0
httpx==0.27.2
huggingface-hub==0.29.0
idna==3.10
importlib-metadata==6.11.0
importlib_metadata==8.6.1
installer==0.7.0
ipykernel==6.29.5
ipython==8.32.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jiter==0.6.1
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
matplotlib-inline==0.1.7
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nanoid==2.0.0
nest_asyncio==1.6.0
networkx==3.4.1
numpy==1.26.1
openai==1.63.2
packaging==24.1
packaging==24.2
parso==0.8.4
pbs-installer==2025.2.12
pendulum==3.0.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
pkginfo==1.12.1.1
platformdirs==4.3.6
poetry-core==2.1.1
poetry==2.1.1
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
questionary==2.1.0
regex==2024.9.11
requests-toolbelt==1.0.0
requests==2.32.3
rsa==4.9
safetensors==0.4.5
setuptools==75.8.0
shellingham==1.5.4
six==1.16.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack_data==0.6.3
starlette==0.45.3
stringcase==1.2.0
sympy==1.13.1
termcolor==2.3.0
tiktoken==0.9.0
time-machine==2.16.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.49.0
trove-classifiers==2025.2.18.16
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.29.2
wcwidth==0.2.13
wheel==0.45.1
xattr==1.1.4
yarl==1.15.4
yaspin==3.1.0
zipp==3.20.2
zipp==3.21.0
zstandard==0.23.0
